requirements.txt
appdirs==1.4.4
imageio==2.9.0
imutils==0.5.3
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pkg-resources==0.0.0
rply==0.7.7
